WIP Refactoring for PR review
