+++ AllReadyApp/Web-App/AllReady.UnitTest/Controllers/ManageControllerPhoneNumberTests.cs
@@ -200,7 +200,7 @@ namespace AllReady.UnitTest.Controllers
             userManagerMock.Setup(x => x.GetUserId(It.IsAny<ClaimsPrincipal>())).Returns(userId);
             var signInManagerMock = MockHelper.CreateSignInManagerMock(userManagerMock);
             var mediator = new Mock<IMediator>();
-            mediator.Setup(x => x.SendAsync(It.IsAny<UserByUserIdQuery>())).ReturnsAsync(new ApplicationUser());
+            mediator.Setup(x => x.SendAsync(It.IsAny<UserByUserIdQuery>())).ReturnsAsync(null);
 
             var controller = new ManageController(userManagerMock.Object, signInManagerMock.Object, mediator.Object);
             controller.SetFakeUser(userId);
